Make sure we keep a stack of active window IDs.
In case Opera closes a window calling the "GetActiveWindow" scope command
may cause it to return the recently closed window Id.
This will be fixed in Opera by forcing Opera to send a new ActivateWindow event when
a window is closed. However that leaves a race condition between the messages, so a
window stack is needed.