Version OperaDriver with output from `git describe`
The version is retreived from Git using `git describe`. This gives
a unique identifier for the current commit, and so the code that
generated a given jar can be found. The version is generated from
the last annotated tag, plus the number of revisions since, plus
a short hash of the commit.

A new (pretty versioned) release can be made by creating an
annotated tag on master.

OperaDriver will now try and load a VERSION file from the
classpath, and will output the value found within as its version
number.