Adds captureOperaIdle
Sometimes when executing a command OperaIdle events will fire
before the response is received for the sent command. This results
in missing the Idle events, and later probably hitting a timeout.

This patch adds captureOperaIdle, which when called will enable
the storing of OperaIdle events until waitForOperaIdle is called.
At that point if OperaIdle events have been received it will return
immediately.