Fix for SPARTAN-777
We never got the idle event in some cases. The timeout was set to 10
seconds so we should not really wait longer than that. If we got any
other event than idle during those 10 seconds we would reset the
timeout, And in the failing example case we got a lot of window loaded
events that continuously reset the idle timeout and thus we never
finished.

This patch keeps track of the time spent waiting for the idle event
such that we in total never wait longer than the initial timeout.