(On behalf of danielb) Fix for SpartanRunner not exiting (SPARTAN-1026)
In certain situations the thread watching output from the launcher
process hangs on stream.read().

I tried various options to kill the thread as calling interrupt() and
to close the stream, using close() but the thread kept living.

However it's not vital for this thread to shut down cleanly so I made
it a daemon thread, thus main() will be exited even if it's still
alive.