Simplified ignore annotation by defaulting to no products/platforms in Ignore
While we typically would want to specify the default to be any product
or any platform, we need to perform two checks for this annotation, both
of which are mutually exclusive.

This means that if one of the checks for platform or product is true,
the other one will be skipped.  If we check for ANY product, it will
always find something to match.