Fixing OperaProfile to work with RemoteWebDriver
Currently fromJson is never called, because the base64-encoded JSON
string in {opera.profile: "foobar="} is a String, so is treated like a
file path by OperaSettings.PROFILE.sanitize.  Instead, encode the
base64-encoded string as a value in a map, so we can distinguish between
it and a file path.

Also, when using the base64-encoded string, currently we create a
TemporaryFileSystem, delete it, and seem to try to access the deleted
file somewhere; instead, just read the file once in to a buffer, and
manipulate that as needed.  Note: I'm assuming that the preferences file
is always stored as UTF-8, as is implied by the comment in
OperaFilePreferences.

And a little tidying up here and there.

Tested by running `ant test` - the same failures exist as in a clean
client, see issue 76.  Added a test that profiles can be decoded
properly.  Manually tested that a profile can be properly loaded through
a RemoteWebDriver server.

Note that this commit technically breaks wire-compatibility from
previously released versions, but the previous versions suffered from a
bug meaning that this never actually worked, so no clients should be
broken by this change (that weren't broken already).