OPDRV-185: Fatal error when evaluating in a non-existent runtime
When attempting to evaluate JS using a runtime that does not exist
using ecmascript service, OperaDriver fatally dies and is not able to
recover.  Using the ecmascript-debugger service (used by Opera 11.64)
this is not an issue.  We can safely ignore a situation when the
runtime doesn't exist because we are always a subject to what the
browser might do.

This patch addresses the issue with using a non-existent runtime when
using the ecmascript service as well as ecmascript-debugger.