OPDRV-188: Selftest API refactor
OperaDriver#selftest() no longer accepts a list of modules which
selftests to run.  It now accepts exactly one module instead.  The
proposed downstream fix for this is to iterate over the list of
modules you wish to test and call this method individually.

The reasoning behind this change is that the old API for running
selftests allowed you to specify a list of modules to run and a
general, overall timeout.  This was not very useful since a large list
of modules would require some heuristic to find an appropriate overall
timeout, and did not allow you to specify the timeout of an individual
module.