Reverted use of getSystemMultiClickTimeout()
The problem was that the call Toolkit.getDefaultToolkit() was very
slow. The time taken varied wildly but on the Spartan slaves I saw
from the fastest 3s and up to as much as 55s! This is an unacceptable
overhead so setting this back to the previous default of 640ms.

The reason is probably this, quote from the java doc:
"Also loads additional classes into the VM..."