OPDRV-189: Improving performance of the WM's window close code
SpartanRunner regression testing showed an added 114 ms overhead in
WindowManager#closeAllWindows() that stemmed from
OperaIntervals#WINDOW_CLOSE_SLEEP.

WindowManager#closeAllWindows() deviates slightly from the Scope
implementation by allowing windows to be closed using Opera's action
system.  The downside of the action system is that there is no
callback, and consequently no way for us to know if a window was
successfully closed.  This is done for product compatibility reasons
as certain products have issues closing dialogues.

WindowManager#closeWindow(i) now has an assertion check for whether
the attached browser service supports closing windows.  If such
support does not exist, it will throw an
UnsupportedOperationException.

Furthermore it will do an implicit wait-style check using a callable
for verifying whether the window requested to be closed is removed
from the internal caches.  This is done though the Scope callback
OnWindowClosed, registered in ScopeEventHandler.

Finally, the patch introduces a variation of Selenium's Wait
(FluentWait) that instead of using Guava's Predicate and Function uses
Callable.  This allows us to perform a wait without any arguments.

Each CallbackWait instance defines the maximum amount of time to wait
for a condition, as well as the frequency with which to check the
condition.  Furthermore, the user may configure the wait to ignore
specific types of exceptions whilst waiting, such as
NoSuchElementException when searching for an element on the page.