Retry and sleep in Eval according to documentation
The documentation of SCRIPT_RETRY_INTERVAL say that it's an interval
not a number of retries. So it looked like the value was 50 ms, but
the loop used that as the number of times to retry the eval. For now I
hardcoded this to 5. I have no idea what a "good" value should be, and
it should not be hardcoded. I'll let the review discuss that part.

Secondly the documentation for sleepDuration say "Defines for how long
to sleep when retrying to execute a JS command. It will be linearly
increased using OperaIntervals.SCRIPT_RETRY_INTERVAL." But that was
not the case, instead it added itself in each iteration. So in total
the sleep duration was set to 50 ms, then doubled itself 50 times.
The 50th sleep would then be 1.78 billion years long :)