Do not try to eval if scripting is disabled
If scripting is disabled in Opera calling eval would throw an
exception after retrying and sleeping a couple of times. Retrying and
sleeping is pointless if we are sure eval would fail anyway so
checking this and returning null in that case. Thus we will in
SpartanRunner also get the same behaviour as with the old driver where
eval simply returned null.