Add proxy support
This patch adds proxy support to OperaDriver.  It exposes an
OperaDriver-specific API OperaDriver#proxy() for accessing the proxy
configuration during runtime.

The capabilities parser in OperaSettings has changed slightly to not
assume an "opera." prefix to any capabilities so that we can also
parse the proxy JSON object sent over the JSON wire protocol.

The proxy object is sanitized in OperaSettings, like all other
capabilities and settings, and the output from
OperaSettings#getProxy() should be safe.

OperaProxy is essentially a wrapper around OperaScopePreferences for
allowing easier access to the proxy configuration.  It also has
OperaProxy#parse() which parses a Selenium Proxy instance.  This will
make some clever deductions on which preferences to switch on and off.