Fix possible exception recursion in gc()
If an exception is thrown in gc() it gets caught by a generic catcher
in start(), which calls quit().  In quit(), however, gc() is called
again.  If gc() fails here we don't really care.

Since the gc() logic is completely different from the shutdown
algorithm elsewhere in quit() it needs to be handled in a separate
try-catch block so that the correct steps are taken for quitting Opera
and shuttiing down the Scope connection.