Nullpointer check in EcmascriptService
We have a number of times (although quite rare and hard to repro) seen
NullPointerExceptions in findRuntime(). So adding a check here. The
doc of getPointer(String xpath) say:

"If the xpath matches no properties in the graph, the pointer will be null."

This is similar to the issue discussed in OPDRV-202.