Rely on RemoteWebElement#equals(obj) to do element equality comparison
We use our unique hash code for this OperaWebElement, constructed from
the DOM object ID and the current runtime ID provided by the WM, to
tell the RemoteWebElement implementation who we are.  This is done so
that we can bypass the problem of two elements gotten by separate
queries which return true for #equals(), but which used to return
unsimilar #hashCode()'s.