Fall back to quit Opera with OperaRunner if all else fails
ScopeServices#quit(OperaRunner) doesn't call OperaRunner#stopOpera().
In fact, that is only called if ScopeServices.quit(…) would throw an
exception.

This introduces a fallback in finally that should ensure Opera is
stopped through the runner if all else failed.  If this too fails
there is really nothing we can do about it, and so it is fine for this
code to reside in a finally block.

Addresses same problem as 7ff60f79bbf394a536990de317933fc2004d5924.