Fix OperaArguments.remove
The current implementation was based on the objects and as we had no equals()
implementation it failed to remove. I let eclise generate the functions as a
quick fix to make it work. But this should be cleaned up and probably solved
differently as it's all just a workaround for moving -fullscreen last on mac.