fix for SPARTAN-1492: Update SpartanRunner to use the new...
This is the webdriver changes for the Scope selftest version 1.0 to 2.0
upgrade. The main change is that the selftest results no longer are
reported as strings, but Selftest.SelftestResult objects, and that the
granularity of the testing has become finer. Groups and exclude filters
can now be used, in addition to modules as before.